refactor: useAddOrUpdateEvent 훅으로 분리
